requirements.txt
tkinterdnd2==0.3.0
pdfminer.six==20221105
python-docx==0.8.11
pyinstaller==5.13.0